Move traits directly in self_intersections.h and add parallelism
